表达式中混用有符号和无符号的告警修改

+++ services/audio_service/common/src/volume_tools.cpp
@@ -277,7 +277,7 @@ static void CountU8Volume(const BufferDesc &buffer, AudioChannel channel, Channe
     }
     // Calculate the average value
     for (size_t index = 0; index < channel; index++) {
-        volMaps.volStart[index] /= frameSize;
+        volMaps.volStart[index] /=static_cast<int32_t>(frameSize);
     }
     return;
 }
@@ -310,7 +310,7 @@ static void CountS16Volume(const BufferDesc &buffer, AudioChannel channel, Chann
     }
     // Calculate the average value
     for (size_t index = 0; index < channel; index++) {
-        volMaps.volStart[index] /= frameSize;
+        volMaps.volStart[index] /= static_cast<int32_t>(frameSize);
     }
     return;
 }
@@ -344,7 +344,7 @@ static void CountS24Volume(const BufferDesc &buffer, AudioChannel channel, Chann
     }
     // Calculate the average value
     for (size_t index = 0; index < channel; index++) {
-        volMaps.volStart[index] /= frameSize;
+        volMaps.volStart[index] /= static_cast<int32_t>(frameSize);
     }
     return;
 }
@@ -377,7 +377,7 @@ static void CountS32Volume(const BufferDesc &buffer, AudioChannel channel, Chann
     }
     // Calculate the average value
     for (size_t index = 0; index < channel; index++) {
-        volSums[index] /= frameSize;
+        volSums[index] /= static_cast<int32_t>(frameSize);
         volMaps.volStart[index] = volSums[index];
     }
     return;
